autoencoder/requirements.txt
torch
torchvision
pandas
pytube
pillow
moviepy